+++ src/main/java/org/realityforge/guiceyloops/server/MockPersistenceTestModule.java
@@ -1,22 +1,17 @@
 package org.realityforge.guiceyloops.server;
 
 import javax.persistence.EntityManager;
-import javax.transaction.UserTransaction;
 
 public class MockPersistenceTestModule
   extends AbstractPersistenceTestModule
 {
   private final boolean _bindWithoutName;
   private final String _persistenceUnit;
-  private final boolean _registerUserTransaction;
 
-  public MockPersistenceTestModule( final String persistenceUnit,
-                                    final boolean bindWithoutName,
-                                    final boolean registerUserTransaction )
+  public MockPersistenceTestModule( final String persistenceUnit, final boolean bindWithoutName )
   {
     _bindWithoutName = bindWithoutName;
     _persistenceUnit = persistenceUnit;
-    _registerUserTransaction = registerUserTransaction;
   }
 
   @Override
@@ -31,15 +26,6 @@ public class MockPersistenceTestModule
     {
       bindMock( EntityManager.class, _persistenceUnit );
     }
-    if ( _registerUserTransaction )
-    {
-      registerUserTransaction();
-    }
-  }
-
-  protected void registerUserTransaction()
-  {
-    bindMock( UserTransaction.class );
   }
 
   @Override